Fixing up functional tests and xoxp decorator
